Commit message: Refactor image handling and improve locale error handling
This commit refactors the way images are handled in ReviewSlide component, importing StaticImageData from next/image and changing the Prop type to StaticImageData[]. This ensures a more precise typing.

In dictionaries.ts, the function getDictionary has been updated to throw an error if the locale is not supported or if there's an error while loading the locale dictionary, to provide more informative error messages.